+++ example/lib/src/loopback.dart
@@ -16,7 +16,6 @@ class Loopback extends StatefulWidget {
 class _LoopbackState extends State<Loopback> {
   List<MediaDeviceInfo>? _mediaDevicesList;
   List<MediaStreamTrack>? _tracks;
-  MediaStreamTrack? _initTrack;
 
   PeerConnection? _pc1;
   RtpTransceiver? _audioTxTr;
@@ -53,7 +52,6 @@ class _LoopbackState extends State<Loopback> {
     if (_inCalling) {
       _hangUp();
     }
-    _initTrack?.dispose();
     _localRenderer.dispose();
     _remoteRenderer.dispose();
   }
@@ -61,15 +59,6 @@ class _LoopbackState extends State<Loopback> {
   void initRenderers() async {
     await _localRenderer.initialize();
     await _remoteRenderer.initialize();
-    // var caps = DeviceConstraints();
-    // caps.audio.mandatory = AudioConstraints();
-    // var tracks = await getUserMedia(caps);
-    // _initTrack = tracks!.firstWhere((track) => track.kind() == MediaKind.audio);
-    // _initTrack!.onAudioLevelChanged((volume) {
-    //   setState(() {
-    //     currentAudioLevel = volume / 100;
-    //   });
-    // });
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.