+++ lib/src/platform/web/media_stream_track.dart
@@ -30,17 +30,6 @@ class WebMediaStreamTrack extends MediaStreamTrack {
   }
 
   @override
-  void onAudioLevelChanged(OnAudioLevelChangedCallback? cb) {
-    // TODO(evdokimovs): implement it using audio level in media-source stats and audio node and AnalyserNode: https://webrtc.github.io/samples/src/content/getusermedia/volume/
-    throw 'onAudioLevelChanged callback currently is not supported on web';
-  }
-
-  @override
-  bool isAudioLevelAvailable() {
-    return false;
-  }
-
-  @override
   bool isEnabled() {
     return jsTrack.enabled ?? false;
   }